add babel 6 support
accord landed babel 6 support in a newer version.

babel 6 is split up into modules, and doesn't bundle any transforms
any more, so I had to add a .babelrc file in the fixtures folder
and add the es2015 preset, as well as replacing the babel dep with
a dependency on babel-core.